Add qr code and once it scan redirect to app
