+++ src/templates/layout/main.hbs
@@ -11,6 +11,9 @@
 
   {{{ body }}}
 
+  <script>
+    _ybg.init()
+  </script>
   <script src="/index.js"></script>
 </body>
 </html>
