+++ src/templates/views/home.hbs
@@ -1,11 +1,12 @@
 <h1>HOME PAGE</h1>
 
-<!-- This span will create the Yoti button -->
-<span data-yoti-application-id="your-app-id" class="register-button">
-  Register runners with Yoti
-</span>
+<div class="button">
+  <span data-yoti-application-id="493d7947-896e-49f6-afd4-e6136b8a9a34">
+    Register runners with Yoti
+  </span>
+</div>
 
 <!-- This script snippet will also be required in your HTML body -->
-<script>
+{{!-- <script>
   _ybg.init()
-</script>
+</script> --}}
