+++ src/templates/views/task-sheet.hbs
@@ -1,1 +1,2 @@
 <h1>TASK SHEET HERE</h1>
+<h2>Hello {{user}}!</h2>
